+++ src/test/java/javax/measure/test/TestUnit.java
@@ -35,10 +35,10 @@ import javax.measure.Dimension;
 import javax.measure.Quantity;
 import javax.measure.Unit;
 import javax.measure.IncommensurableException;
+import javax.measure.Prefix;
 import javax.measure.UnconvertibleException;
 import javax.measure.UnitConverter;
 import javax.measure.quantity.Dimensionless;
-import javax.measure.spi.Prefix;
 import javax.measure.test.unit.BaseUnit;
 import javax.measure.test.unit.MultiplyConverter;
